pyproject.toml
[project]
name = "fast-api-hex"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinicius",email = "vinicius.arruda40@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
